Move all hooks and utils out of shared
