+++ projects/js-packages/charts/src/components/line-chart/line-chart-annotation-label-popover.tsx
@@ -2,7 +2,7 @@ import { __ } from '@wordpress/i18n';
 import clsx from 'clsx';
 import Gridicon from 'gridicons';
 import { useEffect, useId, useRef, useState } from 'react';
-import { isSafari } from '../shared/utils';
+import { isSafari } from '../../utils/is-safari';
 import styles from './line-chart.module.scss';
 import type { ButtonWithPopover, PopoverElement, ToggleEvent } from '../../types';
 import type { FC } from 'react';
