+++ projects/js-packages/charts/src/hooks/index.ts
@@ -2,3 +2,7 @@ export { useDeepMemo } from './use-deep-memo';
 export { useGlobalChartTheme } from './use-global-chart-theme';
 export { useChartMouseHandler } from './use-chart-mouse-handler';
 export { useXYChartTheme } from './use-xychart-theme';
+export { useChartDataTransform } from './use-chart-data-transform';
+export { useChartMargin } from './use-chart-margin';
+export { useElementHeight } from './use-element-height';
+export { useZeroValueDisplay } from './use-zero-value-display';
