+++ projects/js-packages/charts/src/hooks/use-element-height.ts
