+++ projects/js-packages/charts/src/utils/get-longest-tick-width.ts
@@ -23,10 +23,3 @@ export const getLongestTickWidth = < T extends AnyD3Scale >(
 
 	return getStringWidth( longestTick, labelStyle );
 };
-
-export const isSafari = () => {
-	if ( typeof navigator !== 'undefined' && navigator.userAgent ) {
-		return /^((?!chrome|android).)*safari/i.test( navigator.userAgent );
-	}
-	return false;
-};
