+++ projects/js-packages/charts/src/utils/is-safari.ts
@@ -1,0 +1,6 @@
+export const isSafari = () => {
+	if ( typeof navigator !== 'undefined' && navigator.userAgent ) {
+		return /^((?!chrome|android).)*safari/i.test( navigator.userAgent );
+	}
+	return false;
+};
